version 2
Se agregan la API Ciudadano y se modifica la API Administrador, en la url y se corrigen errores

+++ API-Administrador/src/main/java/com/SAFE_Rescue/API_Administrador/service/BomberoService.java
@@ -1,10 +1,15 @@
 package com.SAFE_Rescue.API_Administrador.service;
 
+import com.SAFE_Rescue.API_Administrador.modelo.Credencial;
+import com.SAFE_Rescue.API_Administrador.modelo.Rol;
 import com.SAFE_Rescue.API_Administrador.repository.BomberoRepository;
 import com.SAFE_Rescue.API_Administrador.modelo.Bombero;
+import com.SAFE_Rescue.API_Administrador.repository.CredencialRepository;
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -17,6 +22,12 @@ public class BomberoService {
     @Autowired
     private BomberoRepository bomberoRepository;
 
+    @Autowired
+    private CredencialService credencialService;
+
+    @Autowired
+    private CredencialRepository credencialRepository;
+
     public List<Bombero> findAll(){
         return bomberoRepository.findAll();
     }
@@ -27,12 +38,21 @@ public class BomberoService {
 
     public Bombero save(Bombero bombero) {
         try {
+            Credencial credencial = bombero.getCredencial();
+
             validarBombero(bombero);
 
-            return bomberoRepository.save(bombero);
-        } catch (Exception e) {
+            Credencial guardadaCredencial = credencialService.save(credencial);
 
+            bombero.setCredencial(guardadaCredencial);
+
+            return bomberoRepository.save(bombero);
+        } catch (DataIntegrityViolationException e) {
+            throw new RuntimeException("Error: el correo de la credencial ya está en uso.");
+        } catch (EntityNotFoundException e) {
             throw new RuntimeException("Error al guardar el bombero: " + e.getMessage());
+        } catch (Exception e) {
+            throw new RuntimeException("Error inesperado: " + e.getMessage());
         }
     }
 
@@ -48,17 +68,6 @@ public class BomberoService {
                     throw new RuntimeException("El valor nombre excede máximo de caracteres (50)");
                 }
                 antiguoBombero.setNombre(bombero.getNombre());
-            }
-
-            if (bombero.getCorreo() != null) {
-                if (bomberoRepository.existsByCorreo(bombero.getCorreo())) {
-                    throw new RuntimeException("El Correo ya existe");
-                }else{
-                    if (bombero.getCorreo().length() > 80) {
-                        throw new RuntimeException("El valor correo excede máximo de caracteres (80)");
-                    }
-                    antiguoBombero.setCorreo(bombero.getCorreo());
-                }
             }
 
             if (bombero.getTelefono() != null) {
@@ -118,6 +127,10 @@ public class BomberoService {
 
     public void delete(long id){
         try {
+            if (!bomberoRepository.existsById(id)) {
+                throw new NoSuchElementException("Bombero no encontrado");
+            }
+
             bomberoRepository.deleteById(id);
 
         } catch (Exception e) {
@@ -129,10 +142,6 @@ public class BomberoService {
 
         if (bomberoRepository.existsByRun(bombero.getRun())) {
             throw new RuntimeException("El RUN ya existe");
-        }
-
-        if (bomberoRepository.existsByCorreo(bombero.getCorreo())) {
-            throw new RuntimeException("El Correo ya existe");
         }
 
         if (bomberoRepository.existsByTelefono(bombero.getTelefono())) {
@@ -159,14 +168,21 @@ public class BomberoService {
             throw new RuntimeException("El valor a_materno excede máximo de caracteres (50)");
         }
 
-        if (bombero.getCorreo().length() > 80) {
-            throw new RuntimeException("El valor correo excede máximo de caracteres (80)");
-        }
-
         if (String.valueOf(bombero.getTelefono()).length()> 9) {
             throw new RuntimeException("El valor telefono excede máximo de caracteres (9)");
         }
 
+    }
+
+    public void asignarCredencial(long bomberoId, long credencialId) {
+        Bombero bombero = bomberoRepository.findById(bomberoId)
+                .orElseThrow(() -> new RuntimeException("Bombero no encontrado"));
+
+        Credencial credencial = credencialRepository.findById(credencialId)
+                .orElseThrow(() -> new RuntimeException("Credencial no encontrada"));
+
+        bombero.setCredencial(credencial);
+        bomberoRepository.save(bombero);
     }
 
 }
